refactor : 코드 리펙토링 유저 매니저 킥보드 매니저 분리
